Reorganized command files structure
