+++ src/commands.js
@@ -1,8 +1,8 @@
-const initCommand = require('./init/init.command')
-const runCommand = require('./run/run.command')
-const deployCommand = require('./deploy/deploy.command')
-const scriptsCommand = require('./scripts.command')
-const modulesCommand = require('./modules.command')
+const initCommand = require('./commands/init/init.command')
+const runCommand = require('./commands/run.command')
+const deployCommand = require('./commands/deploy.command')
+const scriptsCommand = require('./commands/scripts.command')
+const modulesCommand = require('./commands/modules.command')
 
 module.exports = exports = {
   initCommand,
