+++ src/commands/deploy.command.js
