+++ src/commands/init/init.command.js
@@ -2,13 +2,13 @@ const fs = require('fs')
 const path = require('path')
 const inquirer = require('inquirer')
 
-const { FsUtils } = require('../shared/fs.utils')
-const { LogUtils } = require('../shared/log.utils')
-const { GitUtils } = require('../shared/git.utils')
+const { FsUtils } = require('../../shared/fs.utils')
+const { LogUtils } = require('../../shared/log.utils')
+const { GitUtils } = require('../../shared/git.utils')
 
 const { InitStopedByUser } = require('./init-stop-by-user.error')
-const { FolderNotEmpty } = require('../shared/folder-not-empty.error')
-const { FileNotDirectory } = require('../shared/file-not-directory.error')
+const { FolderNotEmpty } = require('../../shared/folder-not-empty.error')
+const { FileNotDirectory } = require('../../shared/file-not-directory.error')
 
 const URL_REGEX = /^((?:(https?):\/\/)?((?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.)(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.)(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9]))|(?:(?:(?:\w+\.){1,2}[\w]{2,3})))(?::(\d+))?((?:\/[\w]+)*)(?:\/|(\/[\w]+\.[\w]{3,4})|(\?(?:([\w]+=[\w]+)&)*([\w]+=[\w]+))?|\?(?:(wsdl|wadl))))$/
 const PATH_REGEX = /^(?:(?:~|\.{1,2})?\/)+(?:[a-zA-Z\.\-\_]+\/?)*$/
